+++ jsolex-core/src/main/java/module-info.java
@@ -54,6 +54,7 @@ module me.champeau.a4j.jsolex.core {
     exports me.champeau.a4j.jsolex.processing.sun.crop;
     exports me.champeau.a4j.jsolex.processing.util;
     exports me.champeau.a4j.jsolex.processing.sun.workflow;
+    exports me.champeau.a4j.jsolex.processing.sun.align;
     exports me.champeau.a4j.jsolex.processing.sun.detection;
     exports me.champeau.a4j.jsolex.processing.expr.stacking;
     exports me.champeau.a4j.jsolex.processing.util.spectrosolhub;